Commit 'rote_pop_trends_JAGS.R’
This script specifies a hierarchical trend model that quantifies patterns in the population dynamics of Roseate Terns. The model is akin to the hierarchical models that are used to quantify trend estimates for Breeding Bird Survey data, but I have specified it to better describe the behavior of a colonial population. The basic structure of the tern model is a Poisson generalized linear model with year effects and random intercepts for region. The specification for year effects includes a parameter for first-order temporal autocorrelation. The trend for each island by is determined by a regression equation that includes a fixed effect for the population-wide trend, a random effect for region, and a random effect for island. This script also contains code to 1) obtain posterior distributions for key population measures, including the overall trend line, the percent annual change in population size, annual variation, and region/site effects; and 2) quantify trends in the distribution of the population over time using an index of “dominance” (Simpson’s diversity index).